+++ scripts/rollback/rollback-2.11.0.sh
@@ -1,0 +1,151 @@
+#!/bin/bash
+# timestamp: 1623315032655
+
+npm dist-tag add @midwayjs/egg-layer@2.10.19 latest
+npm dist-tag add @midwayjs/egg-layer@2.11.0 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.10.19 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.11.0 latest
+npm dist-tag add @midwayjs/express-layer@2.10.14 latest
+npm dist-tag add @midwayjs/express-layer@2.11.0 beta
+tnpm dist-tag add @midwayjs/express-layer@2.10.14 latest
+tnpm dist-tag add @midwayjs/express-layer@2.11.0 latest
+npm dist-tag add @midwayjs/koa-layer@2.10.14 latest
+npm dist-tag add @midwayjs/koa-layer@2.11.0 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.10.14 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.11.0 latest
+npm dist-tag add @midwayjs/serverless-app@2.10.17 latest
+npm dist-tag add @midwayjs/serverless-app@2.11.0 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.10.17 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.11.0 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.11.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.10.14 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.11.0 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.11.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.10.14 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.11.0 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.11.0 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.10.14 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.11.0 latest
+npm dist-tag add @midwayjs/static-layer@2.10.14 latest
+npm dist-tag add @midwayjs/static-layer@2.11.0 beta
+tnpm dist-tag add @midwayjs/static-layer@2.10.14 latest
+tnpm dist-tag add @midwayjs/static-layer@2.11.0 latest
+npm dist-tag add @midwayjs/bootstrap@2.10.18 latest
+npm dist-tag add @midwayjs/bootstrap@2.11.0 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.10.18 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.11.0 latest
+npm dist-tag add @midwayjs/cache@2.10.18 latest
+npm dist-tag add @midwayjs/cache@2.11.0 beta
+tnpm dist-tag add @midwayjs/cache@2.10.18 latest
+tnpm dist-tag add @midwayjs/cache@2.11.0 latest
+npm dist-tag add @midwayjs/consul@2.10.18 latest
+npm dist-tag add @midwayjs/consul@2.11.0 beta
+tnpm dist-tag add @midwayjs/consul@2.10.18 latest
+tnpm dist-tag add @midwayjs/consul@2.11.0 latest
+npm dist-tag add @midwayjs/core@2.10.18 latest
+npm dist-tag add @midwayjs/core@2.11.0 beta
+tnpm dist-tag add @midwayjs/core@2.10.18 latest
+tnpm dist-tag add @midwayjs/core@2.11.0 latest
+npm dist-tag add @midwayjs/decorator@2.10.18 latest
+npm dist-tag add @midwayjs/decorator@2.11.0 beta
+tnpm dist-tag add @midwayjs/decorator@2.10.18 latest
+tnpm dist-tag add @midwayjs/decorator@2.11.0 latest
+npm dist-tag add @midwayjs/faas@2.10.18 latest
+npm dist-tag add @midwayjs/faas@2.11.0 beta
+tnpm dist-tag add @midwayjs/faas@2.10.18 latest
+tnpm dist-tag add @midwayjs/faas@2.11.0 latest
+npm dist-tag add @midwayjs/grpc@2.10.18 latest
+npm dist-tag add @midwayjs/grpc@2.11.0 beta
+tnpm dist-tag add @midwayjs/grpc@2.10.18 latest
+tnpm dist-tag add @midwayjs/grpc@2.11.0 latest
+npm dist-tag add @midwayjs/logger@2.10.18 latest
+npm dist-tag add @midwayjs/logger@2.11.0 beta
+tnpm dist-tag add @midwayjs/logger@2.10.18 latest
+tnpm dist-tag add @midwayjs/logger@2.11.0 latest
+npm dist-tag add midway-schedule@2.10.18 latest
+npm dist-tag add midway-schedule@2.11.0 beta
+tnpm dist-tag add midway-schedule@2.10.18 latest
+tnpm dist-tag add midway-schedule@2.11.0 latest
+npm dist-tag add midway@2.10.19 latest
+npm dist-tag add midway@2.11.0 beta
+tnpm dist-tag add midway@2.10.19 latest
+tnpm dist-tag add midway@2.11.0 latest
+npm dist-tag add @midwayjs/mock@2.10.18 latest
+npm dist-tag add @midwayjs/mock@2.11.0 beta
+tnpm dist-tag add @midwayjs/mock@2.10.18 latest
+tnpm dist-tag add @midwayjs/mock@2.11.0 latest
+npm dist-tag add @midwayjs/prometheus@2.10.18 latest
+npm dist-tag add @midwayjs/prometheus@2.11.0 beta
+tnpm dist-tag add @midwayjs/prometheus@2.10.18 latest
+tnpm dist-tag add @midwayjs/prometheus@2.11.0 latest
+npm dist-tag add @midwayjs/rabbitmq@2.10.18 latest
+npm dist-tag add @midwayjs/rabbitmq@2.11.0 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.10.18 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.11.0 latest
+npm dist-tag add @midwayjs/socketio@2.10.18 latest
+npm dist-tag add @midwayjs/socketio@2.11.0 beta
+tnpm dist-tag add @midwayjs/socketio@2.10.18 latest
+tnpm dist-tag add @midwayjs/socketio@2.11.0 latest
+npm dist-tag add @midwayjs/task@2.10.19 latest
+npm dist-tag add @midwayjs/task@2.11.0 beta
+tnpm dist-tag add @midwayjs/task@2.10.19 latest
+tnpm dist-tag add @midwayjs/task@2.11.0 latest
+npm dist-tag add @midwayjs/typegoose@2.10.18 latest
+npm dist-tag add @midwayjs/typegoose@2.11.0 beta
+tnpm dist-tag add @midwayjs/typegoose@2.10.18 latest
+tnpm dist-tag add @midwayjs/typegoose@2.11.0 latest
+npm dist-tag add @midwayjs/version@2.10.19 latest
+npm dist-tag add @midwayjs/version@2.11.0 beta
+tnpm dist-tag add @midwayjs/version@2.10.19 latest
+tnpm dist-tag add @midwayjs/version@2.11.0 latest
+npm dist-tag add @midwayjs/express@2.10.18 latest
+npm dist-tag add @midwayjs/express@2.11.0 beta
+tnpm dist-tag add @midwayjs/express@2.10.18 latest
+tnpm dist-tag add @midwayjs/express@2.11.0 latest
+npm dist-tag add @midwayjs/koa@2.10.18 latest
+npm dist-tag add @midwayjs/koa@2.11.0 beta
+tnpm dist-tag add @midwayjs/koa@2.10.18 latest
+tnpm dist-tag add @midwayjs/koa@2.11.0 latest
+npm dist-tag add @midwayjs/web@2.10.19 latest
+npm dist-tag add @midwayjs/web@2.11.0 beta
+tnpm dist-tag add @midwayjs/web@2.10.19 latest
+tnpm dist-tag add @midwayjs/web@2.11.0 latest
+npm dist-tag add @midwayjs/ws@2.10.14 latest
+npm dist-tag add @midwayjs/ws@2.11.0 beta
+tnpm dist-tag add @midwayjs/ws@2.10.14 latest
+tnpm dist-tag add @midwayjs/ws@2.11.0 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.10.19 => 2.11.0
+#  - @midwayjs/express-layer: 2.10.14 => 2.11.0
+#  - @midwayjs/koa-layer: 2.10.14 => 2.11.0
+#  - @midwayjs/serverless-app: 2.10.17 => 2.11.0
+#  - @midwayjs/serverless-fc-starter: 2.10.14 => 2.11.0
+#  - @midwayjs/serverless-fc-trigger: 2.10.14 => 2.11.0
+#  - @midwayjs/serverless-scf-trigger: 2.10.14 => 2.11.0
+#  - @midwayjs/static-layer: 2.10.14 => 2.11.0
+#  - @midwayjs/bootstrap: 2.10.18 => 2.11.0
+#  - @midwayjs/cache: 2.10.18 => 2.11.0
+#  - @midwayjs/consul: 2.10.18 => 2.11.0
+#  - @midwayjs/core: 2.10.18 => 2.11.0
+#  - @midwayjs/decorator: 2.10.18 => 2.11.0
+#  - @midwayjs/faas: 2.10.18 => 2.11.0
+#  - @midwayjs/grpc: 2.10.18 => 2.11.0
+#  - @midwayjs/logger: 2.10.18 => 2.11.0
+#  - midway-schedule: 2.10.18 => 2.11.0
+#  - midway: 2.10.19 => 2.11.0
+#  - @midwayjs/mock: 2.10.18 => 2.11.0
+#  - @midwayjs/prometheus: 2.10.18 => 2.11.0
+#  - @midwayjs/rabbitmq: 2.10.18 => 2.11.0
+#  - @midwayjs/socketio: 2.10.18 => 2.11.0
+#  - @midwayjs/task: 2.10.19 => 2.11.0
+#  - @midwayjs/typegoose: 2.10.18 => 2.11.0
+#  - @midwayjs/version: 2.10.19 => 2.11.0
+#  - @midwayjs/express: 2.10.18 => 2.11.0
+#  - @midwayjs/koa: 2.10.18 => 2.11.0
+#  - @midwayjs/web: 2.10.19 => 2.11.0
+#  - @midwayjs/ws: 2.10.14 => 2.11.0